Add read and delete routes
