+++ src/config/database.js
@@ -2,7 +2,7 @@ require("dotenv/config");
 
 module.exports = {
   development: {
-    dialect: 'postgres',
+    dialect: "postgres",
     username: process.env.POSTGRES_USER,
     password: process.env.POSTGRES_PASSWORD,
     database: process.env.POSTGRES_DB,
@@ -29,7 +29,7 @@ module.exports = {
     database: process.env.PROD_DB_NAME,
     host: process.env.PROD_DB_HOSTNAME,
     port: process.env.PROD_DB_PORT,
-    dialect: 'postgres',
+    dialect: "postgres",
     dialectOptions: {
       bigNumberStrings: true
     }
