+++ src/routes/index.js
@@ -3,11 +3,13 @@ import AuthRoutes from "./authRoutes";
 import ShiftRoutes from "./shiftRoutes";
 import UserRoutes from "./userRoutes";
 import RoleRoutes from "./roleRoutes";
+import CompanyRoutes from "./companyRoutes";
 
 const router = Router()
   .use("/auth", AuthRoutes)
   .use("/shifts", ShiftRoutes)
   .use("/users", UserRoutes)
-  .use("/roles", RoleRoutes);
+  .use("/roles", RoleRoutes)
+  .use("/companies", CompanyRoutes);
 
 export default router;
